move conftest contents into a pytest plugin
we will switch to a entrypoint at a later point in time
